Reorganise network code to prevent ConcurrentModificationException.
This can happen when the server-side packet handling attempts to access
an entity while the EntityTracker is performing
updateTimeLightAndEntities which can result in a concurrent modification
error.

The solution for this is to schedule the packet processing task to be
done in the main thread, after updateTimeLightAndEntities has completed.

I've split the packet into Server/Client packets simply for ease of use,
as it's extremely confusing when one packet is being usesd in both
directions. Additionally, as the processing takes place outside of the
network thread, the information is now manually transmitted back to the
client.

I may have introduced other issues, I'm uncertain, but I will be testing
it in my modpack to make sure that the crahses no longer occur.